requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-bootstrap-form==3.4
django-debug-toolbar==3.6.0
djangorestframework==3.13.1
fontawesomefree==6.2.0
Pillow==9.2.0
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
